Refactored to account for more warnings

+++ Sequential_Archive/Packer.cpp
@@ -151,7 +151,7 @@ bool Packer::Pack_File_With_Buffers(QFile &file, QFile &sourceFile, qint64 fileS
 
 QByteArray Packer::Get_Byte_Array_From_qint64(qint64 number) {
     QByteArray byteArray(8, 0x00);
-    byteArray.data()[0] = static_cast<char>(static_cast<unsigned char>(((number&0xFF00000000000000)>>56)&0xFF));
+    byteArray.data()[0] = static_cast<char>(static_cast<unsigned char>(((static_cast<quint64>(number)&0xFF00000000000000)>>56)&0xFF));
     byteArray.data()[1] = static_cast<char>(static_cast<unsigned char>((number&0x00FF000000000000)>>48));
     byteArray.data()[2] = static_cast<char>(static_cast<unsigned char>((number&0x0000FF0000000000)>>40));
     byteArray.data()[3] = static_cast<char>(static_cast<unsigned char>((number&0x000000FF00000000)>>32));
@@ -164,7 +164,7 @@ QByteArray Packer::Get_Byte_Array_From_qint64(qint64 number) {
 
 QByteArray Packer::Get_Byte_Array_From_int(int number) {
     QByteArray byteArray(4, 0x00);
-    byteArray.data()[0] = static_cast<char>(static_cast<unsigned char>(((number&0xFF000000)>>24)&0xFF));
+    byteArray.data()[0] = static_cast<char>(static_cast<unsigned char>(((static_cast<unsigned int>(number)&0xFF000000)>>24)&0xFF));
     byteArray.data()[1] = static_cast<char>(static_cast<unsigned char>((number&0x00FF0000)>>16));
     byteArray.data()[2] = static_cast<char>(static_cast<unsigned char>((number&0x0000FF00)>>8));
     byteArray.data()[3] = static_cast<char>(static_cast<unsigned char>((number&0x000000FF)));
